1. set required attribute in BaseStreamForm class 2. don't interfere with field-group choices 3. set required attribute in field-groups for editing submissions
