+++ hypha/apply/stream_forms/models.py
@@ -1,6 +1,9 @@
 # Credit to https://github.com/BertrandBordage for initial implementation
 import copy
+
 from collections import OrderedDict
+
+from django import forms
 
 from wagtail.contrib.forms.models import AbstractForm
 
@@ -46,11 +49,12 @@ class BaseStreamForm:
     def get_defined_fields(self):
         return self.form_fields
 
-    def get_form_fields(self, draft=False):
+    def get_form_fields(self, draft=False, posted_field_values={}):
         form_fields = OrderedDict()
         field_blocks = self.get_defined_fields()
         group_counter = 1
         is_in_group = False
+        sub_children_visible = False
         for struct_child in field_blocks:
             block = struct_child.block
             struct_value = struct_child.value
@@ -65,6 +69,7 @@ class BaseStreamForm:
                     field_from_block.grouper_for = group_counter + 1
                     group_counter += 1
                     is_in_group = True
+                    sub_children_visible = posted_field_values.get(struct_child.id) == field_from_block.choices[0][0]
                 if isinstance(block, TextFieldBlock):
                     field_from_block.word_limit = struct_value.get('word_limit')
                 if isinstance(block, MultiInputCharFieldBlock):
@@ -85,6 +90,10 @@ class BaseStreamForm:
                             field_from_block.initial = None
                         field_from_block = copy.copy(field_from_block)
                 else:
+                    if is_in_group and not isinstance(field_from_block, forms.ChoiceField):
+                        field_from_block.required_when_visible = field_from_block.required
+                        field_from_block.required = field_from_block.required & sub_children_visible
+                        field_from_block.visible = sub_children_visible
                     form_fields[struct_child.id] = field_from_block
             elif isinstance(block, GroupToggleEndBlock):
                 # Group toogle end block is used only to group fields and not used in actual form.
@@ -97,8 +106,8 @@ class BaseStreamForm:
 
         return form_fields
 
-    def get_form_class(self, draft=False):
-        return type('WagtailStreamForm', (self.submission_form_class,), self.get_form_fields(draft))
+    def get_form_class(self, draft=False, posted_field_values={}):
+        return type('WagtailStreamForm', (self.submission_form_class,), self.get_form_fields(draft, posted_field_values))
 
 
 class AbstractStreamForm(BaseStreamForm, AbstractForm):
